Separate SelectUnit from Degree Units page into its own component
